refactor: decouple gas limit schedule logging
